Refactor to separate packets and handlers
