+++ src/game_server/handlers/guid.rs
