+++ src/game_server/handlers/mod.rs
@@ -1,0 +1,15 @@
+pub mod character;
+pub mod chat;
+pub mod command;
+pub mod guid;
+pub mod housing;
+pub mod inventory;
+pub mod item;
+pub mod lock_enforcer;
+pub mod login;
+pub mod mount;
+pub mod reference_data;
+pub mod test_data;
+pub mod time;
+pub mod unique_guid;
+pub mod zone;
