+++ src/game_server/packets/purchase.rs
@@ -1,8 +1,9 @@
 use std::io::Write;
 
-use crate::game_server::game_packet::{GamePacket, OpCode};
 use byteorder::{LittleEndian, WriteBytesExt};
 use packet_serialize::{DeserializePacket, SerializePacket, SerializePacketError};
+
+use super::{GamePacket, OpCode};
 
 #[derive(Copy, Clone, Debug)]
 pub enum PurchaseOpCode {
